requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
python-multipart
passlib
python-dotenv
firebase-admin
psycopg2-binary
itsdangerous
